+++ .github/workflows/basic.yml
@@ -34,7 +34,7 @@ jobs:
     - name: Build
       run: make build
     - name: InstUt
-      run: go test -v $(go list -e ./pkg/... | grep -v /rules/) -coverprofile=coverage.txt -covermode=atomic
+      run: go test -v $(go list -e ./pkg/... | grep -v /rules/ | grep github.com/alibaba/opentelemetry-go-auto-instrumentation/pkg/) -coverprofile=coverage.txt -covermode=atomic
     - name: Upload coverage reports to Codecov
       uses: codecov/codecov-action@v4.0.1
       with:
